refactor(plans): migrate DeletePlanDialog to new CentralizedDialog hook
Replaces the deprecated `WarningDialog`-based `DeletePlanDialog` ref API
with a `useDeletePlanDialog` hook using the NiceModal `CentralizedDialog`
system. The file is renamed to `useDeletePlan.tsx` to avoid the
`*Dialog*` restricted-import lint pattern at consumer sites.
